Fetch build & manifest repo from Github
Move out of a local directory and start using Github for the
manifest repo.

Right now I'm using the same repo for both the build scripts and the
manifest. There is probably a better way to do this, but I'm not
familiar enough with repo to know how. Can be fixed up later.

Signed-off-by: Grant Likely <grant.likely@arm.com>