Rockpro64: add target to flash SD
This is a *terrible* hack, but it gets a target added for flashing the
firmware to an SD. What is needed is a way to abstract platform specific
flash targets so that each target can do what it needs.

In the case of the rk3399 platforms, it would be best to create a GPT on
the SD card with a couple of protected partitions for the firmware
images. I'll look into doing that later.

Signed-off-by: Grant Likely <grant.likely@arm.com>