Refactor to use u-boot defconfig to get defaults
U-Boot defconfig already knows a lot about the platform. Use the
defconfig data to know what the SoC is and start with sane defaults.

This currently works by running the U-Boot defconfig step with a
:=$(shell) statement before executing any make rules. It is a little
ugly to do it this way, and it may be better to recursively call Make
so that the output of make <target>_defconfig is executed in a proper
rule.
Signed-off-by: Grant Likely <grant.likely@arm.com>