Clean up specifying build directory
Instead of setting the build directory unconditionally (which doesn't
entirely work anyway; U-Boot handles it well, but TFA and DT-rebasing do
not), by default let each project use it's own working directory.
However, if BUILD_BASE is set, then use that to create build directories
for the projects that support it.

Signed-off-by: Grant Likely <grant.likely@arm.com>