Refactor into a single makefile for core build
This patch adopts the Linux & U-boot kbuild style top level Makefile
that includes the generated config so that decisions can be made from
the specific config. The core of this change is to pull in the %config
parsing rules for u-boot/Makefile and process them separately from the
main build rules. This allows *config rules to change the U-Boot
configuration, and the import the CONFIG_* values into make so that
TFA_PLAT and other variables can be automatically determined.

Signed-off-by: Grant Likely <grant.likely@arm.com>