Map clean targets onto u-boot clean targets
Make the 'clean', 'mrproper' & 'distclean' targets match the u-boot
targets of the same name. Also use the 'clean' targets for TFA and the
devicetree-rebasing tree as well as make all clean targets callable when
unconfigured.

For U-Boot, all targets are made callable when not configured because
this tool is entirely driven by the U-Boot config.

Signed-off-by: Grant Likely <grant.likely@arm.com>