Rename edk2 targets to be consistent and not match real files
edk2/BaseTools is a real directory, so using that as a target means the
target won't get built. Also the stmm target is built from the edk2
tree, so it makes sense to prefix the target with 'edk2-'. Change both
targets to be prefixed with 'edk2-' which should not conflict with the
directory name and can be set up as .PHONY targets.

Signed-off-by: Grant Likely <grant.likely@arm.com>